GO, KEGG, and Reactome over-representation analysis (ORA)

Perform GO, KEGG, and Reactome ORA for all proteins (genes) in a glyexp::experiment().

This function uses the "protein" column in the variable information tibble. Protein identifiers should be UniProt accessions.

Usage

enrich_go(exp, return_raw = FALSE, ...)

enrich_kegg(exp, return_raw = FALSE, ...)

enrich_reactome(exp, return_raw = FALSE, ...)

Arguments

exp

A glyexp::experiment() object.

return_raw

A logical value indicating whether to return raw clusterProfiler enrichResult objects.

...

Additional arguments passed to clusterProfiler::enrichGO(), clusterProfiler::enrichKEGG(), or ReactomePA::enrichPathway().

Value

A tibble of GO, KEGG, or Reactome enrichment results (when return_raw = FALSE), or raw clusterProfiler enrichResult objects (when return_raw = TRUE).

Required packages

This function requires the following packages to be installed:

clusterProfiler for enrichment analysis

ReactomePA for Reactome pathway analysis

org.Hs.eg.db for human gene annotation (GO analysis only)

See also