Reactome pathway over-representation analysis (ORA)

Perform Reactome pathway ORA for protein UniProt accessions using ReactomePA::enrichPathway().

gly_enrich_reactome() accepts a glyexp::experiment() and extracts protein information from the "protein" column in the variable information tibble.

gly_enrich_reactome_() accepts a character vector of UniProt IDs.

As ReactomePA::enrichPathway() only accepts Entrez IDs, the UniProt IDs will be first transformed into Entrez IDs with clusterProfiler::bitr().

Usage

gly_enrich_reactome(
  exp,
  add_info = TRUE,
  orgdb = "org.Hs.eg.db",
  organism = "human",
  universe = NULL,
  p_adj_method = "BH",
  p_cutoff = 0.05,
  q_cutoff = 0.2
)

gly_enrich_reactome_(
  proteins,
  orgdb = "org.Hs.eg.db",
  organism = "human",
  universe = NULL,
  p_adj_method = "BH",
  p_cutoff = 0.05,
  q_cutoff = 0.2
)

Arguments

exp

(Only for gly_enrich_reactome()) A glyexp::experiment() object.

add_info

A logical value. This parameter is included for API consistency but has no effect since enrichment results do not contain variable or sample columns. Only applicable to top-level APIs.

orgdb

Passed to OrgDb of clusterProfiler::bitr(). Organism database name (e.g., "org.Hs.eg.db" for human). Defaults to "org.Hs.eg.db".

organism

Passed to organism of ReactomePA::enrichPathway(). Species name (e.g., "human", "mouse", "rat"). Defaults to "human".

universe

Background genes. If a character vector, directly passed to universe of ReactomePA::enrichPathway(). You can also provide a glyexp::experiment() object with "glycoproteomics" type. In this case all detected proteins in this experiment will be extracted and passed to universe of ReactomePA::enrichPathway().

p_adj_method

Passed to pAdjustMethod of ReactomePA::enrichPathway().

p_cutoff

Passed to pvalueCutoff of ReactomePA::enrichPathway().

q_cutoff

Passed to qvalueCutoff of ReactomePA::enrichPathway().

proteins

(Only for gly_enrich_reactome_()) A character vector of UniProt accession IDs.

Value

A list with three elements:

tidy_result: A tibble with enrichment results containing the following columns:

id: Term ID (e.g., R-HSA-XXXXXX)

description: Term description

gene_ratio: Ratio of genes in the term to total genes in the input

bg_ratio: Ratio of genes in the term to total genes in the background

p_val: Raw p-value from hypergeometric test

p_adj: Adjusted p-value

q_value: Q-value (FDR)

gene_id: Gene IDs in the term (separated by "/")

count: Number of genes in the term

raw_result: The raw ReactomePA enrichPathway result object

meta_data (only for gly_enrich_reactome()): A list containing metadata from the input experiment The list has classes glystats_reactome_ora_res and glystats_res.

Required packages

These functions require the following packages to be installed:

clusterProfiler for ID conversion

ReactomePA for enrichment analysis

org.Hs.eg.db for human gene annotation or other OrgDb packages

See also